[ADD] l10n_pe: adding missing taxes
Adding a new set of taxes to cover the cases that we are missing, and
adding the codes to identify the taxes.

Based on the information we find in http://cpe.sunat.gob.pe/sites/default/files/inline-files/AjustesValidacionesCPEv20180726.zip
We are adding the next data to create the taxes.

[ADD] l10n_pe: add new taxes according to catalog 05

Based on the update of 24/06/2019 from http://cpe.sunat.gob.pe/node/88
we are validating the catalog 5 with the most relevant taxes from
catalog 5.

- IVAP: Tax applied for rice at 4%, it does not require IGV applied to
this product.
- ISC: Selective tax for consumers, it is applied to an specific group
of products, usually are products that are nocive to the health of the
consumer an are meant to drive the consumer away from this product.
- ICBPER: Special tax that will be charged to plastic bags starting in
august of 2019.
- EXO: Exonerated taxes are those that depends on special conditions to
allow the IGV(base tax) to don't be applied.
- INA: Unaffected taxes are for those products that will never be taxed
with IGV or any tax.