[ADD] l10n_pe: adding missing taxes
Adding a new set of taxes to cover the cases that we are missing, and
adding the codes to identify the taxes.

Based on the information we find in http://cpe.sunat.gob.pe/sites/default/files/inline-files/AjustesValidacionesCPEv20180726.zip
We are adding the next data to create the taxes.