[FIX] l10n_pe: Fix data prefix and views
Fixing the prefix of the model res.district to l10n.pe,district, in
order to let know the user or future developers that is a model that
works for Peru.
Updating some fields that were proposed for the version 12 and now those
fields are differing in master version.
Also adding some access depending on the user group.