[IMP] l10n_pe: Renaming files and changing licenses
Renaming the model l10n.pe.district to l10n_pe.res.city.district to indicate
that it is a child of city, also changing the name of its file and the CSV
file, to match the name of the model.
Renaming the name of the field "l10n_pe_edi_district" to "l10n_pe_district,
due to this field is now a Peruvian field and not an EDI field.
Removing all the Vauxoo's copyright because all these changes are going to be
part of odoo's base code, also changing the license "LGPL-3.0" to Odoo
licenses.