[FIX] account: Update invoice_date / date in Customer Invoices at Duplication
When an Customer Invoice is duplicated it shall bear Today's Date for reference purposes
in the Invoice to be posted. It must not copy the same old date from the original document.
This is useful when duplicating Multi-currency invoices

If the Document got duplicated yesterday and It happens to be posted tomorrow.
It shall computed the values at tomorrows rate.
Not that of the date when it was first computed.