[REF] base: don't delete records related if PG error is raised
In following lines:

* https://github.com/odoo/odoo/blob/bffe0d00b798672aec7624cb0a5ea1fb1200cc99/odoo/addons/base/wizard/base_partner_merge.py#L200
* https://github.com/odoo/odoo/blob/bffe0d00b798672aec7624cb0a5ea1fb1200cc99/odoo/addons/base/wizard/base_partner_merge.py#L175

The unlinking is done due to the all postgres errors raised in the updating
of foreign keys and references, but this is so dangerous if the partner is not deleted
as the out-of-box code, instead we are only archive the partners
and also the deletion could cause concurrence errors that cause merges
in incorrect states and documents related with partners incorrect.

Then we return if the rollback was done to handling the post-processes.