[IMP] web: Handle the margins of computeFloatOuterWidthWithMargins
The problem is that the "+" dropdown appears on the header even when there is enough space for other menu items.

The function initAutoMoreMenu() is used to adapt the header layout so that when there is no space for an element in the
header a "+" dropdown appears and will contain that element.

This issue is not in v13 where we had the same header and this one collapses as we expect. This is because in the old
version the code calculates if there is a margin (right or left) before calculating the max width
(see:https://github.com/Vauxoo/odoo/blob/13.0/addons/web/static/src/js/core/dom.js#L518) with the computeFloatOuterWidthWithMargins
function.

In v15 the computeFloatOuterWidthWithMargins function evaluates if there are margins with a condition
(https://github.com/Vauxoo/odoo/blob/15.0/addons/web/static/src/legacy/js/core/menu.js#L162) but this condition even if
is true or false always keep the true because the considerAutoMargins that is always true and the negation of the
expression doesn't allow a correct functionality as in v13. Where the only value to take is the margins.

We change the condition with one as in v13 that only evaluates the margin and we keep the direction of the language
in the account so the function is still the same.