[REF] pos_gift_card: Gift Card split by taxes
Allow splitting a gift card when 2 or more products with different taxes
are in the same order, so each group of taxes has a line with the Gift
Card assigned with the corresponding taxes.

For that, the validation of using the same gift card in the order
"This gift card is already applied" was removed to allow the creation of
multiple lines with the same gift card at the same time, but the gift
card will only be used once by the order since when is used a second
time the previous lines will be removed.

Also, the computing of the `balance` in the gift card was fixed to allow
using the `price_subtotal_incl` instead of `price_unit` since when a tax
without the option "Included in Price" (`price_include`) set as `True`
the balance was not being calculated taking into account those taxes.