[REF] sale_timesheet: Reproduce concurrency issue creating task timesheet
We have observed concurrency issue
when attempting to create two or more records at the same time
by different users

If the company grows these such errors will be more common
Specially if there are more users creating timesheet records on the same task

Should we look for a way to fix it?
Or
Should we request to users only retry and retry?