[FIX] mail: lock mail.mail record before sending email
In this line:

https://github.com/odoo/odoo/blob/c58a34294112b9d156fd563c207aaaacb1a90eb5/addons/mail/models/mail_mail.py#L730

The changes applied in mail.mail sending the email
are applied, if there is another process at the same
time updating the same record, that is going
to raise a concurrence error. The problem with this
is the email went out already, but in Odoo, it keeps
as `outgoing`, causing multiple emails went out
for same record in Odoo.

This commit adds a `flush_recordset` before to
send the email, if there is another process trying
to update the same record, it will be blocked and
it must wait to sending email process finishes.
The mail record will be updated as `sent`.