[IMP] http_routing: Avoid returning verbose errors
Returning verbose errors presents a vulnerability to subsequent attacks
by sharing information that could be used to find and exploit
vulnerabilites. To fix this, the traceback sent to the template for the
error page now displays a hash instead. This hash will match a hash that
will accompany the error within the log, so that errors raised can still
be debugged