[IMP] auth_totp,base,web: New validation messages
This allows changing strings created from code that cannot be changed
using i18n or i18n_extra in IRC due to the new form of saving
translations.

Changed the message displayed when the user has too many login failures,
the message when the user use a wrong login or password in web module
and the message that appears when the verification through 2fa is not
successful in auth_totp module.

Related to:
https://www.vauxoo.com/web#id=75454&cids=1&action=470&model=project.task&view_type=form