[FIX] website: Unused Attribute Warning in Render of `list_hybrid` template
When searching on the Website site, using the main search on the navbar,
it opens a list view with the results (`website.list_hybrid`) which
raises the warning on the logs (2 times):
"Unknown directives or unused attributes: {'t-key'} in
website.list_hybrid"

This happens after the attribute `t-key` was added to the template [1]
because the template is only used in QWeb. The validation for them
doesn't include the `t-key` [2] as one of the "iter_directives" nor has
a `_compile_directive_*` method to check and remove it from the
validation as it's done with the `t-as` and `t-foreach`.

This also causes the raise of the warnings on tours that use the tour
method `searchProduct` (of the module `website_sale`) because it uses
the first input with the name of search and happens to be the search on
the navbar.

[1]: https://github.com/odoo/odoo/commit/7b1d82aa
[2]: https://github.com/odoo/odoo/blob/f52cfb09/odoo/addons/base/models/ir_qweb.py#L1400